+++ src/yamk/functions.py
@@ -39,6 +39,15 @@ class Exists(Function):
         return self.base_dir.joinpath(path).exists()
 
 
+class Name(Function):
+    name = "name"
+
+    def __call__(self, path):
+        if isinstance(path, list):
+            return [self(file) for file in path]
+        return self.base_dir.joinpath(path).name
+
+
 class Stem(Function):
     name = "stem"